Upgraded the testing further to avoid manually updates
I changed some things in test.py and the individual tests.
In test.py, in no circumstance do you need to add the test modules
manually. It finds them itself based on their name. I made some of the
tests more unittest-oriented as well but a lot still to be done on that.

bbelson@rainier.princeton.edu