Speed up freqresp for transfer functions by an order of magnitude
When determining the frequency response, the list of frequencies
is now stored as a numpy array, instead of a list, and evaluated
all at once, instead of using `map`.  This is much faster.

The improved performance can be measured using a script
`bench/time_freqresp.py`.  On my machine, I get a speedup of
more than a factor of 20 (from 3.9 seconds to 0.18 seconds) for
a 10th-order transfer function.