Quiet down noisy MATLAB tests
Fixes #31, suppressing warnings for converting from MIMO to SISO
if the user has explicitly specified the input to use, in
`step_response`, `initial_response`, and `impulse_response`.

If a test should give a warning (such as an impulse response for
a system with a direct feedthrough term), test that a warning was issued,
but suppress the warning from the output of the test.