Fix some MATLAB tests
Fixes #30, suppressing some known warnings for systems with direct
feedthrough terms.  Also removes a duplicate test in
`test_control_matlab`, and fixes some previously broken, skipped
tests:

  test_convert_MIMO_to_SISO
  test_impulse

This update also removes a previous test to make sure warning
messages are issued: this turns out to be unreliable, because
the default behavior is for warnings to be issued only once, so
depending on the order in which tests are executed, the warning
may or may not be raised.