Fix dependencies and mocks in sphinx documentation
There was an error in the way the sphinx documentation was
configured that prevented it from building correctly, now that
absolute imports (e.g., "from control import lti") have been changed
to relative imports (e.g., "from . import lti").

In particular, the directory "root/control" should not be added
to sys.path.  Because this directory contains an __init__.py file,
it is a Python package, and packages should never be added to the
path, or lots of strange things happen, particularly with relative
imports.  Instead, the directory "root" should be added to the
path.  This change was made in the conf.py file.

Also, the mocks for modules were fixed so that the documentation
builds even if numpy, scipy, etc are not installed.

Conflicts:
	doc/conf.py