Fix setup.py to work with runtests.py
Previously, at least on my machine (Mac with Anaconda),
running runtests.py gave a build error when setup.py tried to
build a local copy of the package.

It turns out this was because we were importing setuptools,
which caused problems, because the rest of the setup.py script
(which was taken from numpy) assumes we are using the numpy
version of setup(), in numpy.distutils.core

This commit removes the import of setuptools and makes setup.py
almost identical to that used in numpy.