Simplify setup.py and automatic version generation
There was a lot of unnecessarily complex logic in setup.py, taken from
numpy, about whether to use setuptools or distutils.  In addition, the
procedure for generating the version information was only partially
automatic (one still needed to change the version manually in setup.py),
and was somewhat unreliable (see issue #37).

This commit simplifies setup.py, always using setuptools, as recommended
in the Python Packaging Guide:
    http://python-packaging-user-guide.readthedocs.org/en/latest/current.html
The version information is now generated directly from tags in the git
repository.  Now, *before* running setup.py, one runs
    python make_version.py
and this generates a file with the version information.  This is copied
from binstar (https://github.com/Binstar/binstar) and seems to work well.