Fix bug in stability_margins
Fixes #58.  The documentation is also updated to specify the units of
the inputs and outputs.  Also, previously there was an optional
argument `deg` which the docstring claimed was used to specify the
units of the input/output phases, but this was never actually used in
the implementation.  This commit removes that optional argument.