Reorganize documentation
Previously, the modules were documented on separate pages (in the html
version), with one html page per module.  This commit changes the
organization of the documentation, so that there are concise tables of
functions (or classes), plus one html page per function or class.  This
is consistent with how numpy, scipy, and other projects handle their
documentation.