Move documentation into doc/
The tables of functions/classes were previously defined in
control/__init__.py.  This commit moves those to doc/control.rst.  This
is consistent with what is done in scikit-learn, for example, and seems
cleaner to me, as most of the .rst is kept in the doc/ directory, rather
than the docstrings in __init__.py.  It also shortens a too-lengthy
docstring for the main `control` module.