Clean up imports
Define __all__ variables in the modules so that
    from .<module> import *
imports only the functions/classes we use.  This is probably a little
better than listing them explicitly in __init__.py, since then if we add
a new function to a module, we only need to change the __all__ variable
in the same module, rather than edit __init__.py.  Furthermore it avoids
duplicating some of these listings in matlab/__init__.py.