Implement dcgain outside of matlab module
The dcgain function is now implemented as a method of the StateSpace and
TransferFunction classes.  There is also a routine lti.dcgain(sys) that
simply calls sys.dcgain().

Note that the DC gain is now returned as a numpy array (not a matrix),
and furthermore, single dimensions are removed with np.squeeze.  This is
a change from the previous implementation in matlab.dcgain().  The new
convention makes it easier to deal with SISO systems (a common case),
since instead of referring to a gain as g[0][0], you can refer to it
simply as g.  This seems preferable to me than returning a
np.matrix.  (The new convention is also more consistent with Matlab.)