Always return phase in radians in bode_plot
Previously, bode_plot returned phase and magnitude in different
units (dB/absolute magnitude, radians/degrees) depending on the values
of boolean flags.  This was inconsistent with the documentation and the
tests.  This is now made consistent, so that bode_plot always returns
absolute magnitude and phase in radians, and the documentation is
consistent with the behavior.