Update evalfr usage for consistency with MATLAB
Renamed evalfr() in the FRD, StateSpace and TransferFunction classes to
_evalfr() and put a deprecation warning for use of the evalfr() method.

Changed calls to evalfr() in frdata.py and margins.py to use _evalfr()

Added unit tests for the the new methods + warnings

These changes eliminate the inconsistency in the argument form between the
(MATLAB compatible) evalfr() function, which takes a complex argument, and
the evalfr() methods for LTI objects, which took a real number as argument
(representing the frequency).  The _evalfr() method retains the original
functionality, since it is used for computing stability margins and for
converting objects to FRD form.