Fix: remove deprecated subplot() use
Sub-axes are created with specific axis labels, which are strings;
the Bode plot uses 'pycontrol-bode-{mag,phs}', and the gang-of-four
plot uses 'pycontrol-gof-{t,s,ps,cs}' to identify its plots.

These axes are then used, if possible; if not, the figure is cleared
and the set of axes created.