Fix namespace and logic errrors in statesp.freqresp + unit tests
As pointed out in PR #192, there were some namespace errors in the statesp
module related to checking whether the frequency range as valid for discrete
time systems.  I added a unit test that covered this code (and exposed the
error) and fixed up the namespace problems (as in PR #192).  There was also
a logic error in the way that frequencies were checked and the warning message
referred to the wrong function.