initial implementation of optimization for flatsystems
This commit allows a cost function and constraints to be passed to the
flatsys point_to_point() function and when present will use sp.optimize to
find a trajectory.  Preliminary unit tests, benchmarks, docstrings and
documentation also in place.

Switched the order of arguments in point_to_point() so that Tf (or timepts)
now comes before the initial and final states, consistent with ordering
elsewhere in the package.

Make some small updates to optimal.py docstrings and argument names to make
things consistent with flatsys.