Lint: remove unused variables, or prefix with `_`
Where it's clear that function is called for side effects (e.g., in a
q`with pytest.raises` block), don't assign function output.  Where
it's not clear, e.g., binary ops on LTI objects, call result `_sys` or
similar.  There are plenty of in-between cases: for those I chose
based on understandability.