fix color cycling not working in singular_values_plot
this seems to have ben caused by a name collision between the function scope color variable and the loop-"local" color variable.
the first time _get_color is called it will replace the function level color so in the next iteration a color is explicitly passed to _get_color so it will no longer automatically cycle through colors.